pyproject.toml
[tool.poetry]
name = "semantic-router"
version = "0.0.29"
description = "Super fast semantic router for AI decision making"
authors = [
    "James Briggs <james@aurelio.ai>",
    "Siraj Aizlewood <siraj@aurelio.ai>",
    "Simonas Jakubonis <simonas@aurelio.ai>",
    "Luca Mannini <luca@aurelio.ai>",
    "Bogdan Buduroiu <bogdan@aurelio.ai>",
    "Ismail Ashraq <ashraq@aurelio.ai>",
    "Daniel Griffin <daniel@aurelio.ai>"
]
readme = "README.md"
packages = [{include = "semantic_router"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pydantic = "^2.5.3"
openai = "^1.10.0"
cohere = "^4.32"
mistralai= {version = "^0.0.12", optional = true}
numpy = "^1.25.2"
colorlog = "^6.8.0"
pyyaml = "^6.0.1"
pinecone-text = {version = "^0.7.1", optional = true}
fastembed = {version = "^0.2.4", optional = true, python = "<3.12"}
torch = {version = "^2.1.0", optional = true}
transformers = {version = "^4.36.2", optional = true}
llama-cpp-python = {version = "^0.2.28", optional = true}
colorama = "^0.4.6"
pinecone-client = {version="^3.0.0", optional = true}
regex = "^2023.12.25"
torchvision = { version = "^0.17.0", optional = true}
pillow = { version= "^10.2.0", optional = true}
tiktoken = "^0.6.0"
matplotlib = { version="^3.8.3", optional = true}
qdrant-client = {version="^1.8.0", optional = true}

[tool.poetry.extras]
hybrid = ["pinecone-text"]
fastembed = ["fastembed"]
local = ["torch", "transformers", "llama-cpp-python"]
pinecone = ["pinecone-client"]
vision = ["torch", "torchvision", "transformers", "pillow"]
processing = ["matplotlib"]
mistralai = ["mistralai"]
qdrant = ["qdrant-client"]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"
ruff = "^0.1.5"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.5.0"
mypy = "^1.7.1"
types-pyyaml = "^6.0.12.12"
types-requests = "^2.31.0"
black = "^24.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.per-file-ignores]
"*.ipynb" = ["ALL"]

[tool.ruff]
line-length = 88

[tool.mypy]
ignore_missing_imports = true